Split main and collector files
to better distinguish code from node_exporter
